move services to highlight dir
